Clarified Function Name: renameInitialization
Previously there were two instances of functions called rename, one was
a setup function that called the function which does the actual
renaming. The latter has been kept the same where as the prior is now
known as renameInitialization.

The Console and GUI frontends have been updated to match this change.